requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
requests
python-dotenv
bigframes
google-cloud-aiplatform
matplotlib
seaborn
google.cloud
